Home page reuses BaseAppBar; eliminate HomeAppBar
